Fixed warnings in InventorySystem.

+++ Company Users/Company Users/InventorySystem.cpp
@@ -77,7 +77,7 @@ void Inventory::loadDevices() {
 		getline(ss, field, ','); description = field;
 		int	model_ID = atoi(szmodel.c_str());
 		int	user_ID = atoi(szuser.c_str());
-		float setPrice = atof(szprice.c_str());
+		float setPrice = (float)atof(szprice.c_str());
 		// If the loaded device had bigger id than newid then we need to increase the newid
 		// because we don't want to set this same id to any other device we might create later
 		curid = newid++;
@@ -244,7 +244,7 @@ void Inventory::removeDevice() {
 void Inventory::showDevices() {
 	cout << "  ID     Name " << endl;
 	cout << "============================ " << endl;
-	for (int i = 0; i < deviceList.size(); i++) { //(int)deviceList.size();
+	for (int i = 0; i < (int)deviceList.size(); i++) { //(int)deviceList.size();
 		cout << " [" << deviceList[i].getDevice_ID() << "] - " << deviceList[i].getName() << endl; //displays device ID and device name
 	}
 	cout << "============================ " << endl << endl;
@@ -256,7 +256,7 @@ void Inventory::showModelDevices() {
 
 	cout << "  Model ID     Name " << endl;
 	cout << "============================ " << endl;
-	for (int i = 0; i < deviceList.size(); i++) { //(int)deviceList.size();
+	for (int i = 0; i < (int)deviceList.size(); i++) { //(int)deviceList.size();
 		cout << " [" << deviceList[i].getModel_ID() << "] - " << deviceList[i].getName() << endl; //displays device ID and device name
 	}
 	cout << "============================ " << endl;
@@ -267,7 +267,7 @@ void Inventory::totalValue() {
 
 	float inventoryValue = 0; // sets inv value to 0
 
-	for (int i = 0; i < deviceList.size(); i++) { //loops through so many times how many devices there is
+	for (int i = 0; i < (int)deviceList.size(); i++) { //loops through so many times how many devices there is
 		if (deviceList[i].getPrice()) { //when price is found
 			inventoryValue += deviceList[i].getPrice(); //adds devices value to inventory Value
 		}
